requirements.txt
streamlit
openai
azure.identity
streamlit_chat
azure.core
azure.search.documents